+++ src/main/java/duke/Event.java
@@ -1,30 +1,34 @@
+package duke;
+
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
-class Deadline extends Task {
+class Event extends Task {
 
     private Date date;
     private DateFormat formatter = new SimpleDateFormat("dd/MM/yy h:mm a");
 
-    public Deadline(String description, Date date) {
+
+    public Event(String description, Date date) {
         super(description);
         this.date = date;
     }
 
-    public Deadline(String description, boolean hasCompleted, Date date) {
+    public Event(String description, boolean hasCompleted, Date date) {
         super(description);
         this.status = hasCompleted;
         this.date = date;
+
     }
 
     @Override
     public String toString() {
-        return "[" + Type.D + "]";
+        return "[" + Type.E + "]";
     }
 
     @Override
     String getDescription() {
-        return this.description + " (by: " + formatter.format(date).toString() + ")";
+        return this.description + " (at: " + formatter.format(date).toString() + ")";
     }
 }
