+++ src/main/java/duke/HelpCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.IOException;
 import java.text.ParseException;
