+++ src/main/java/duke/Task.java
@@ -1,6 +1,4 @@
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+package duke;
 
 class Task {
     enum Type {
